Convert all LWRPs to custom resources
Modernizes the cookbook to use custom resources instead of LWRPs.

Includes a few touch-ups to the dokken test suites.
