+++ recipes/users.rb
@@ -40,7 +40,7 @@ node['rabbitmq']['enabled_users'].each do |user|
   user['rights'].each do |r|
     rabbitmq_user "set-perms-#{user['name']}-vhost-#{Array(r['vhost']).join().tr('/', '_')}" do
       user user['name']
-      vhost r['vhost']
+      vhost r['vhost'] if r['vhost']
       permissions "#{r['conf']} #{r['write']} #{r['read']}"
       action :set_permissions
     end
