+++ resources/parameter.rb
@@ -22,10 +22,57 @@
 
 unified_mode true if respond_to?(:unified_mode)
 
-actions :set, :clear, :list
 default_action :set
 
-attribute :parameter, :kind_of => String, :name_attribute => true
-attribute :component, :kind_of => String
-attribute :vhost, :kind_of => String
-attribute :parameters, :kind_of => [Hash, Array], :default => {}
+property :parameter, String, name_property: true
+property :component, String
+property :vhost, String, default: '/'
+property :parameters, [Hash, Array], default: {}
+
+action_class do
+  include RabbitMQ::CoreHelpers
+
+  def parameter_exists?(vhost, name)
+    cmd = 'rabbitmqctl list_parameters -s'
+    cmd += " -p #{Shellwords.escape vhost}" unless vhost.nil?
+    cmd += " |grep '#{name}\\b'"
+
+    cmd = Mixlib::ShellOut.new(cmd, :env => shell_environment)
+    cmd.run_command
+    !cmd.error?
+  end
+end
+
+action :set do
+  cmd = 'rabbitmqctl set_parameter -q'
+  cmd += " -p #{new_resource.vhost}"
+  cmd += " #{new_resource.component}"
+  cmd += " #{new_resource.parameter}"
+
+  cmd += " '"
+  cmd += JSON.dump(new_resource.parameters)
+  cmd += "'"
+
+  parameter = "#{new_resource.component} #{new_resource.parameter}"
+
+  execute "set_parameter #{parameter}" do
+    command cmd
+    environment shell_environment
+    not_if { parameter_exists?(new_resource.vhost, new_resource.parameter) }
+  end
+
+  Chef::Log.info "Done setting RabbitMQ parameter #{parameter}."
+end
+
+action :clear do
+  parameter = "#{new_resource.component} #{new_resource.parameter}"
+
+  cmd = "rabbitmqctl clear_parameter #{parameter} -q -p #{new_resource.vhost}"
+  execute "clear_parameter #{parameter}" do
+    command cmd
+    environment shell_environment
+    only_if { parameter_exists?(new_resource.vhost, new_resource.parameter) }
+  end
+
+  Chef::Log.info "Done clearing RabbitMQ parameter #{parameter}."
+end
